✨ [FEAT] 채팅 읽음 기능 구현

+++ src/main/java/com/boro/domain/chat/dto/response/ChatRoomPreview.java
@@ -11,7 +11,7 @@ public record ChatRoomPreview(
         String profileUrl,
         String lastMessageContent,
         LocalDateTime lastMessageAt,
-        Integer unreadCount,
+        Long unreadCount,
         String postTitle
 ) {
 }
